requirements.txt
altgraph==0.16.1
anvil-uplink==0.3.34
arcade==2.0.9
asgihandler==0.5.4
asgiref==3.7.2
attrs==19.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2019.9.11
channels==4.0.0
chardet==3.0.4
colorama==0.4.6
coverage==7.2.7
cPython==0.0.6
crispy-bootstrap4==2022.1
curses-check-for-keypress==0.3
curses-menu==0.5.0
curses-ui==0.0.2
curses-util==0.0.12
cursgame==0.2.0
cycler==0.10.0
Cython==3.0.2
Deprecated==1.2.14
distlib==0.3.7
Django==3.2.20
django-bootstrap4==23.1
django-cleanup==8.0.0
django-crispy-forms==2.0
django-environ==0.10.0
django-model-utils==4.3.1
django-quiz-app==0.5.1
django-rest-framework==0.1.0
django-session-timeout==0.1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
dnspython==2.3.0
docutils==0.15.2
eli5==0.10.1
ensure==0.8.2
environ==1.0
exceptiongroup==1.1.3
filelock==3.12.2
future==0.17.1
gopay==1.3.1
graphviz==0.13
h5py==2.10.0
humanize==4.6.0
idna==2.8
importlib-metadata==6.7.0
inflect==6.0.4
iniconfig==2.0.0
Jinja2==3.1.2
joblib==0.13.2
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
llvmlite==0.39.1
MarkupSafe==2.1.3
matplotlib==3.1.1
mysql-connector-python==8.0.23
mysqlclient==1.4.6
numba==0.56.4
numpy==1.21.6
packaging==23.1
pandas==0.25.1
pefile==2019.4.18
Pillow==6.1.0
platformdirs==3.10.0
pluggy==1.2.0
prettytable==0.7.2
protobuf==4.23.4
psycopg2==2.9.6
pyassimp==4.1.4
pydantic==1.10.7
pygame==1.9.6
pyglet==1.4.0b1
pyglet-ffmpeg2==0.1.12
Pygments==2.4.2
PyInstaller==3.5
pymongo==4.5.0
PyMySQL==1.1.0
pyparsing==2.4.2
pypi-sandbox==0.0.0
pyproject-api==1.5.3
pysandbox==66.0.2
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.0
pytz==2019.2
pywin32-ctypes==0.2.0
PyYAML==5.1.2
reportlab==3.5.59
requests==2.22.0
RestrictedPython==6.1
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.11.0
six==1.12.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==5.5.0
table==0.0.4
tabulate==0.8.3
Theano==1.0.4
tomli==2.0.1
typing-extensions==4.7.1
urllib3==1.25.3
virtualenv==20.24.3
windows-curses==1.1
wrapt==1.15.0
ws4py==0.5.1
wxPython==4.0.6
zipp==3.15.0
